Test ear clipping
